requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database drivers
psycopg2-binary==2.9.9
pymongo==4.6.0
asyncpg==0.29.0
motor==3.3.2

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers>=2.2.0,<3.0.0
huggingface-hub>=0.16.0

# AI integration
google-generativeai==0.3.2

# Additional utilities
pandas==2.1.4
numpy==1.25.2
python-jose[cryptography]==3.3.0

# Kafka and Redis (for distributed processing)
aiokafka==0.8.11
redis==5.0.1

# HTTP client for testing
httpx==0.25.2
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

# Kafka integration
aiokafka==0.8.11
redis==5.0.1
